fix: skip recording metrics when no service is matched
If Kong has an error or if Kong is returning
a 404 when no Route is matches, the plugin will
be executed and no service will be matched.

This change avoids recording those metrics and a potential
nil dereference, resulting in noise in logs.

Cases when plugin is executed when Kong has an error
or other cases should be recorded by the plugin
but is out of scope for this fix.