WIP: Export Prometheus metrics from Streaming module
At the moment, Lua code running in the context of HTTP module cannot share
any state with Lua code running in the context of Streaming module
(even through shared memory).

That's why we're forced to have 2 separate endpoints with Prometheus metrics:
one for HTTP module and another for Streaming module.

However, we don't want our users to start depending on it -
from their perspective there should be a single Prometheus endpoint.

To achieve this, Prometheus endpoint of HTTP module will be making
a sub-request to the Prometheus endpoint of Streaming module under the hood.