feat(api) made api.lua compatible with Kong's `next` branch
The way we used to implement custom APIs has changed in the
latest versions of Kong. The `schema` parameter (used for
defining this plugin's API by using an "empty schema trick") is no
longer used nor understood by Kong. If not removed, newer versions
of kong won't be able to properly expose the `/metrics` endpoint.